requirement.txt
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.0.0
streamlit>=1.30.0
Pillow>=9.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
